requirements.txt
ipykernel
camelot-py
PyMuPDF
ghostscript
matplotlib
pandas